Remove 16-bit encodings
The compressed future instructions section caused encoding clashes with the code-size reduction Zce, and caused confusion in the PLCT implementation of the tools, so I've removed them. C.NOT, C.NEG, and C.ZEXT.W are included in Zce already. C.ZEXT.D isn't currently included in Zce as it is considers instructions for RV32/RV64 but does not specifically add instructions for RV128.